requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
psutil==5.9.5
bcrypt==4.0.1
